% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enp.R
\name{enp}
\alias{enp}
\title{The effective number of parties}
\usage{
enp(x)
}
\arguments{
\item{x}{(\code{numeric}). A number vector in the form of a proportion of votes or seats. The vector should range from 0 to 1.}
}
\description{
`enp()` This function calculates the number of effective parties according to the proposition of Laakso and Taagepera (1979).
}
\examples{
\dontrun{
df <- enp(c(0.2, 0.3, 0.5))
}
}
\references{
Laakso, M., & Taagepera, R. (1979). “Effective” Number of Parties: A Measure with Application to West Europe. Comparative Political Studies, 12(1), 3–27. https://doi.org/10.1177/001041407901200101
}
