% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frac.R
\name{fractionalization}
\alias{fractionalization}
\title{Fractionalization Index}
\usage{
fractionalization(x)
}
\arguments{
\item{x}{(\code{numeric}). A number vector in the form of a proportion of votes or seats. The vector should range from 0 to 1.


`fractionalization()` Returns a numerical vector with the estimate of party fractionalization.}
}
\description{
`fractionalization()` This function calculates the rate of electoral fractionalization according to
Douglas Rae's (1971) proposal.
}
\examples{
\dontrun{
df <- fractionalization(c(0.2, 0.3, 0.5))
}
}
\references{
Rae, D. W. (1971). Comment on Wildgen's The Measurement of Hyperfractionalization. Comp. Pol. Stud., 4, 244.
}
