% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hh.R
\name{hh}
\alias{hh}
\title{Herfindahl–Hirschman concentration index}
\usage{
hh(x)
}
\arguments{
\item{x}{(\code{numeric}). A number vector in the form of a proportion of votes or seats. The vector should range from 0 to 1.}
}
\description{
`hh()` This function calculates the Herfindahl–Hirschman concentration index.
}
\examples{
\dontrun{
df <- hh(x)
}
}
