% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiperfrac.R
\name{hiperfrac}
\alias{hiperfrac}
\title{hyperfractionalization index}
\usage{
hiperfrac(x)
}
\arguments{
\item{x}{(\code{numeric}). A number vector in the form of a proportion of votes or seats. The vector should range from 0 to 1.}
}
\description{
`hiperfrac` hyperfractionalization proposed by Kesselman (1966) and Wildgen (1971) is an alternative when the system has many small parties.
}
\details{
This index gives a greater role to small parties compared to the effective number of parties `enp`
}
\examples{
\dontrun{
df <- hiperfrac(x)
}
}
\references{
Kesselman, M. (1966). French Local Politics: A Statistical Examination of Grass Roots Consensus. American Political Science Review, 60(4), 963-973. doi:10.2307/1953769

Wildgen, J. K. (1971). The Measurement of Hyperfractionalization. Comparative Political Studies, 4(2), 233–243. https://doi.org/10.1177/001041407100400205
}
