% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desp_lh.R
\name{lh}
\alias{lh}
\title{Loosemore–Hanby index}
\usage{
lh(x, y)
}
\arguments{
\item{x}{(\code{numeric}). Numeric vector with the vote share of parties}

\item{y}{(\code{numeric}). Numeric veector with the seat shareof parties}
}
\description{
`lh()` This function calculates the electoral disproportionality between votes and
seats as proposed by Loosemore and Hanby.
}
\examples{
lh(x, y)
}
