% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tag.R
\name{Tag}
\alias{Tag}
\title{Imbalance Index}
\usage{
Tag(x)
}
\arguments{
\item{(\code{numeric}).}{A number vector in the form of a proportion of votes or seats. The vector should range from 0 to 1.}
}
\description{
`Tag()`This function calculates inbalance according to the proposition of Taagepera (1979).
}
\examples{
Tag(x)
}
