% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbi.R
\name{tbi}
\alias{tbi}
\title{T Imbalance Index}
\usage{
tbi(x)
}
\arguments{
\item{(\code{numeric}).}{A number vector in the form of a proportion of votes or seats. The vector should range from 0 to 1.}
}
\description{
`tbi()`This function calculates T index of inbalance according to the proposition of Taagepera (1979).
}
\examples{
tbi(x)
}
