% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desp_lsq.R
\name{lsq}
\alias{lsq}
\title{Least squares index}
\usage{
lsq(x, y)
}
\arguments{
\item{x}{(\code{numeric}). Numeric vector with the vote share of parties}

\item{y}{(\code{numeric}). Numeric veector with the seat shareof parties}
}
\description{
`lsq()` This function calculates the electoral disproportionality between votes and
seats by Least squares index method as proposed by Michael Gallagher.
}
\examples{
lsq(x, y)
}
