% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nationalization.R
\name{nationalization}
\alias{nationalization}
\title{Party Nationalization Index}
\usage{
nationalization(x)
}
\arguments{
\item{x}{(\code{numeric}). Numeric vector with the vote share for a party x in each of the electoral districts.}
}
\description{
`nationalization()` This function calculates the party nationalization index as proposed by Jones e Mainwaring (2003).
}
\examples{
\dontrun{
df <- nationalization(x)
}
}
