% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbi.R
\name{tbi}
\alias{tbi}
\title{T Imbalance Index}
\usage{
tbi(x)
}
\arguments{
\item{x}{(\code{numeric}). A number vector in the form of a proportion of votes or seats. The vector should range from 0 to 1.}
}
\description{
`tbi()` This function calculates T index of inbalance according to the proposition of Taagepera (1979).
}
\examples{
\dontrun{
df <- tbi(c(0.2, 0.3, 0.5))
}
}
\references{
Taagepera, R. (1979). Inequality, Concentration, Imbalance. Political Methodology, 6(3), 275–291. http://www.jstor.org/stable/25791080
}
