% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volat.R
\name{volatility}
\alias{volatility}
\title{Electoral Volatility Index}
\usage{
volatility(x, y)
}
\arguments{
\item{x}{(\code{numeric}). Numeric vector with the vote share for a parties in the election x.}

\item{y}{(\code{numeric}). Numeric vector with the vote share for a parties in the election y.}
}
\description{
`volatility()` This function calculates electoral volatility index by Perdersen (1979).
}
\examples{
\dontrun{
df <- volatility(x)
}
}
\references{
Pedersen, M. N. (1979). The dynamics of European party systems: changing patterns of electoral volatility. European journal of political research, 7(1), 1-26.
}
